+++ rollup.config.js
@@ -33,7 +33,7 @@ const shared = {
 			customElement: true
 		}),
 		resolve(),
-		production && terser()
+		terser()
 	]
 }
 
@@ -65,7 +65,7 @@ export default [
 				dev: !production
 			}),
 			resolve(),
-			production && terser()
+			terser()
 		],
 		input: 'src/app.js',
 		output: {