+++ zoo-modules/button-module/button.spec.js
@@ -67,5 +67,38 @@ describe('Zoo button', function() {
 			expect(defaultAttrs.smallSize).to.be.true;
 			expect(defaultAttrs.primaryType).to.be.true;
 		});
+
+		it('should fall back to primary type when deprecated cold is passed', async() => {
+			const result = await page.evaluate(() => {
+				let button = document.createElement('zoo-button');
+				button.type = 'cold';
+				document.body.appendChild(button);
+				const nestedButton = button.shadowRoot.querySelector('button');
+				return nestedButton.classList.contains('primary');
+			});
+			expect(result).to.be.true;
+		});
+
+		it('should fall back to secondary when deprecated hot is passed', async() => {
+			const result = await page.evaluate(() => {
+				let button = document.createElement('zoo-button');
+				button.type = 'hot';
+				document.body.appendChild(button);
+				const nestedButton = button.shadowRoot.querySelector('button');
+				return nestedButton.classList.contains('secondary');
+			});
+			expect(result).to.be.true;
+		});
+
+		it('should create hollow button', async () => {
+			const result = await page.evaluate(() => {
+				let button = document.createElement('zoo-button');
+				button.type = 'hollow';
+				document.body.appendChild(button);
+				const nestedButton = button.shadowRoot.querySelector('button');
+				return nestedButton.classList.contains('hollow');
+			});
+			expect(result).to.be.true;
+		});
 	});
 });