+++ zoo-modules/collapsable-list-module/CollapsableList.svelte
@@ -1,17 +1,15 @@
 <svelte:options tag="zoo-collapsable-list"></svelte:options>
-<div class="box">
-	<ul>
-		{#each items as item, idx}
-			<li class="item" class:active="{_items && _items[idx].active}"> 
-				<span class="header" on:click="{e => handleItemHeaderClick(e, idx)}">
-					{item.header}
-					<svg width="24" height="24" viewBox="0 0 24 24"><path d="M7.41 8.59L12 13.17l4.59-4.58L18 10l-6 6-6-6 1.41-1.41z"/></svg>
-				</span>
-				<slot name="item{idx}"></slot>
-			</li>
-		{/each}
-	</ul>
-</div>
+<ul>
+	{#each items as item, idx}
+		<li class="item" class:active="{_items && _items[idx].active}"> 
+			<span class="header" on:click="{e => handleItemHeaderClick(e, idx)}">
+				{item.header}
+				<svg width="24" height="24" viewBox="0 0 24 24"><path d="M7.41 8.59L12 13.17l4.59-4.58L18 10l-6 6-6-6 1.41-1.41z"/></svg>
+			</span>
+			<slot name="item{idx}"></slot>
+		</li>
+	{/each}
+</ul>
 
 <style type="text/scss">
 	@import "variables";