+++ zoo-modules/grid-module/GridHeader.svelte
@@ -1,7 +1,9 @@
 <svelte:options tag="zoo-grid-header"></svelte:options>
 <div class="box" bind:this={gridHeaderRoot}>
 	<slot></slot>
-	<svg class="arrow" sortstate={sortState} on:click="{() => handleSortClick()}" width="24" height="24" viewBox="0 0 24 24"><path d="M7.41 8.59L12 13.17l4.59-4.58L18 10l-6 6-6-6 1.41-1.41z"/></svg>
+	<svg class="arrow" sortstate={sortState} on:click="{() => handleSortClick()}" width="24" height="24" viewBox="0 0 24 24">
+		<path d="M7.41 8.59L12 13.17l4.59-4.58L18 10l-6 6-6-6 1.41-1.41z"/>
+	</svg>
 </div>
 
 <style type='text/scss'>